moves and renames Bundler and Modulariser - now part of bootstrap as they have no important role in binding
