+++ se.jbee.inject.bootstrap/main/java/se/jbee/inject/bootstrap/BundleBootstrapper.java
@@ -1,22 +1,24 @@
 /*
- *  Copyright (c) 2012-2019, Jan Bernitt 
- *			
+ *  Copyright (c) 2012-2019, Jan Bernitt
+ *
  *  Licensed under the Apache License, Version 2.0, http://www.apache.org/licenses/LICENSE-2.0
  */
-package se.jbee.inject.bind;
+package se.jbee.inject.bootstrap;
+
+import se.jbee.inject.bind.Bundle;
 
 /**
  * Determines all reachable {@link Bundle}s starting from a root {@link Bundle}.
- * 
- * @author Jan Bernitt (jan@jbee.se)
+ *
+ * This is more a utility API for testing.
  */
 @FunctionalInterface
-public interface Bundler {
+public interface BundleBootstrapper {
 
 	/**
 	 * @param root origin of reachable computation
 	 * @return All {@link Bundle}s (their {@link Class}es) that are reachable
 	 *         (installed) when starting from the given root {@link Bundle}.
 	 */
-	Class<? extends Bundle>[] bundle(Class<? extends Bundle> root);
+	Class<? extends Bundle>[] installedBundles(Class<? extends Bundle> root);
 }
